webpack config refactor
