+++ webpack.prod.js
@@ -1,0 +1,20 @@
+var merge = require('webpack-merge');
+
+const MiniCssExtractPlugin = require('mini-css-extract-plugin');
+var OptimizeCssAssetsPlugin = require('optimize-css-assets-webpack-plugin');
+const TerserPlugin = require('terser-webpack-plugin');
+var Visualizer = require('webpack-visualizer-plugin');
+
+var common = require('./webpack.common.js');
+
+module.exports = merge(common, {
+  mode: 'production',
+  optimization: {
+    minimizer: [new TerserPlugin()],
+  },
+  plugins: [
+    new MiniCssExtractPlugin(),
+    new OptimizeCssAssetsPlugin(),
+    new Visualizer({ filename: './statistics.html' }),
+  ],
+});
